create a service common crate
