+++ console-backend/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 
 console-common = { path = "../console-common" }
+service-common = { path = "../service-common" }
 
 actix = "0.10"
 actix-cors = "0.5"
