+++ service-common/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "service-common"
+version = "0.1.0"
+authors = ["Jean-Baptiste Trystram <jbtrystram@redhat.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+#openId
+envconfig = "0.9"
+openid = "0.7"
+log = "0.4"
+
+anyhow = "1"
+reqwest = "0.10"
+failure = "0.1" # required for converting error away from failure
+url = "2"
+pem = "0.8"
+
+# Error
+actix-web = "3"
+serde = "1"
+snafu = "0.6"
