+++ service-common/src/error.rs
